Rename pupilService.js to pupil.service.js
